Temporary fix for maybe<T, U> not compiling when T and U where not trivially destructible.
This was due to the union causing the type to have an implicity deleted destructor. I have the destruction functions defined however I had a lapse in reasoning and forgot destruction order.

The union itself will likely need to be contained in a base type that is inherited from in order to implement the trivial destruction optimizations.