No copy/move types (#77)
* Added type `jc::no_copy_construct` which has its copy constructor deleted. This is really just a building block type.

Added test for it as well.

* Added `jc::no_copy_assign` which has its copy assignment operator deleted.

This is also more of a building block...

Added test for it as well.

* Added `jc::no_copy` which has its copy constructor and copy assignment operator deleted.

This is also a building block but for sure a more useful one.

Added test for it as well.

* Added `jc::no_move_construct` that has its move constructor deleted.

Added test for it as well.

* Added `jc::no_move_assign` that has its move assignment operator deleted.

Added test for it as well.

* Added `jc::no_move` that has its move constructor and move assignment operator deleted.

Added test as well.

MSVC may have an issue with this. If a type inherits from a base type with the base's move constructor deleted. The derived type's move constructor is implicity deleted as well (see https://en.cppreference.com/w/cpp/language/move_constructor). I am uploading this to test if gcc and clang have the same behaviour.

* `std::is_move_constructible` and `std::is_move_assignable` will be true for derived types where the parent's move is deleted!!!! This is wierd behavior as it will return false for that parent type.

Modified how I define the `no_move` type to make sure the test passes.

* Redefined `jc::no_copy` to match `jc::no_move`

* Added `jc::no_copy_or_move` that deletes the copy/move constructors and assignment operators.

Added test for it as well.

* Moved copy_or_move related tests into its own test file